Windows compiler error.
Former-commit-id: 9858169ef1e044aca85dff4140f8c0e1575dfd18

+++ source/service/visualization/testing/cxTestRenderSpeed.cpp
@@ -153,7 +153,7 @@ void TestRenderSpeed::addViewsToLayout(QLayout* layout)
 
 void TestRenderSpeed::addViewsToGridLayout(QGridLayout* layout)
 {
-	int squareNumViews = sqrt(this->getNumViews());
+	int squareNumViews = sqrt((double)this->getNumViews());
 	for (int i = 0; i < this->getNumViews(); i++)
 		layout->addWidget(mViews[i], i / squareNumViews, i % squareNumViews);
 }
